+++ src/Janus.Cli/Program.cs
@@ -8,6 +8,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Janus.Cli.Bootstrap;
 using Janus.Cli.Configuration;
+using Janus.Cli.Erasures;
 using Janus.Cli.Rotation;
 using Janus.Core;
 
@@ -45,6 +46,7 @@ internal static class Program
             [RotateKeyEncryptionKeyCommand.Name] = RotateKeyEncryptionKeyCommand.RunAsync,
             [RotateFingerprintKeyCommand.Name] = RotateFingerprintKeyCommand.RunAsync,
             [ConfigureCommand.Name] = ConfigureCommand.RunAsync,
+            [ReplayErasuresCommand.Name] = ReplayErasuresCommand.RunAsync,
         };
 
     /// <summary>